students-projects-code/team3-cross-lingual-sentiment/team3-GenAI-Team3/datafetch/requirements.txt
requests
beautifulsoup4
lxml
newspaper3k
fake_useragent
tqdm
langdetect
selenium